src/requirements.txt
Flask==2.2.3
uWSGI==2.0.21
PyJWT==2.6.0
mysql-connector-python==8.0.30
Jinja2==3.1.2
uuid==1.30
boto3==1.26.99
python-dotenv==1.0.0
DateTime==5.0
requests==2.28.2
pytest==7.2.2